[tuke @ 4995e71273f94a7d71a6fa5151297ddbb03341a7] Magic dynamic subclassing implemented.
However, turns out it suffers from a fatal flaw. The method resolution
order, __mro__ causes the Element subclass to be checked when looking for
attributes, so for instance the _save() related unittests fail as the
save() function is getting the wrapped object rather than the Element it
is expecting to deal with.