a.add(Element) now works.
Had to do some fancy special-casing in the wrapper code to achieve this. Now
that the Element.__new__ constructor returns a wrapped element add(elem) was
failing as elem would be wrapped.

Currently the unit tests are failing for various reasons, but a short script,
test_element_add, has been included in this commit for testing.