Fixed GC problems with Source.
This uncovered some other bugs. First of all __dict__ had to be defined for
Source instances, not exactly sure how that was working before. Secondly due to
the more convoluted gc reference graph a serious bug in how Wrapped_dealloc
removes self from the wrapped_cache was uncovered.

Plenty of tests failing, but it's a step in the right direction.